+++ backend/src/user/user.controller.ts
@@ -1,11 +1,4 @@
-import {
-    Controller,
-    Post,
-    Body,
-    Param,
-    Get,
-    UploadedFile,
-} from '@nestjs/common';
+import { Controller, Post, Body, Param, Get } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { AddRoleDto } from './dto/add-role.dto';
@@ -20,8 +13,8 @@ export class UserController {
     }
 
     @Post()
-    create(@UploadedFile() avatar: File, @Body() createUserDto: CreateUserDto) {
-        return this.usersService.create(createUserDto, avatar);
+    create(@Body() createUserDto: CreateUserDto) {
+        return this.usersService.create(createUserDto);
     }
 
     @Post('/:id/role')